@@ -1,7 +1,7 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>disable</Nullable>
   </PropertyGroup>
@@ -15,17 +15,18 @@
   </ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
-		<PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.6.0" />
-		<PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.1" />
+		<PackageReference Include="AutoMapper" Version="12.0.1" />
+		<PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
+		<PackageReference Include="MediatR" Version="12.0.1" />
 	</ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\corePackages\Core.Application\Core.Application.csproj" />
-    <ProjectReference Include="..\..\..\corePackages\Core.CrossCuttingConcers\Core.CrossCuttingConcerns.csproj" />
-    <ProjectReference Include="..\..\..\corePackages\Core.Mailing\Core.Mailing.csproj" />
-    <ProjectReference Include="..\..\..\corePackages\Core.Persistence\Core.Persistence.csproj" />
-    <ProjectReference Include="..\Domain\Domain.csproj" />
+	  <ProjectReference Include="..\..\..\corePackages\Core.Application\Core.Application.csproj" />
+	  <ProjectReference Include="..\..\..\corePackages\Core.CrossCuttingConcers\Core.CrossCuttingConcerns.csproj" />
+	  <ProjectReference Include="..\..\..\corePackages\Core.ElasticSearch\Core.ElasticSearch.csproj" />
+	  <ProjectReference Include="..\..\..\corePackages\Core.Mailing\Core.Mailing.csproj" />
+	  <ProjectReference Include="..\..\..\corePackages\Core.Persistence\Core.Persistence.csproj" />
+	  <ProjectReference Include="..\Domain\Domain.csproj" />
   </ItemGroup>
 
 </Project>
