@@ -2,8 +2,8 @@
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 using System.Text;
+using System.Text.Json;
 
 namespace Core.Application.Pipelines.Caching;
 
@@ -11,47 +11,95 @@ public class CachingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, 
     where TRequest : IRequest<TResponse>, ICachableRequest
 {
     private readonly IDistributedCache _cache;
+    private readonly CacheSettings _cacheSettings;
     private readonly ILogger<CachingBehavior<TRequest, TResponse>> _logger;
 
-    private readonly CacheSettings _cacheSettings;
-
-    public CachingBehavior(IDistributedCache cache, ILogger<CachingBehavior<TRequest, TResponse>> logger,
-                           IConfiguration configuration)
+    public CachingBehavior(IDistributedCache cache, IConfiguration configuration, ILogger<CachingBehavior<TRequest, TResponse>> logger)
     {
         _cache = cache;
         _logger = logger;
-        _cacheSettings = configuration.GetSection("CacheSettings").Get<CacheSettings>();
+        _cacheSettings = configuration.GetSection("CacheSettings").Get<CacheSettings>() ?? throw new InvalidOperationException();
     }
-
-    public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken,
-                                        RequestHandlerDelegate<TResponse> next)
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
     {
+        if (request.BypassCache)
+            return await next();
+
         TResponse response;
-        if (request.BypassCache) return await next();
-
-        async Task<TResponse> GetResponseAndAddToCache()
-        {
-            response = await next();
-            TimeSpan? slidingExpiration =
-                request.SlidingExpiration ?? TimeSpan.FromDays(_cacheSettings.SlidingExpiration);
-            DistributedCacheEntryOptions cacheOptions = new() { SlidingExpiration = slidingExpiration };
-            byte[] serializeData = Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(response));
-            await _cache.SetAsync(request.CacheKey, serializeData, cacheOptions, cancellationToken);
-            return response;
-        }
-
         byte[]? cachedResponse = await _cache.GetAsync(request.CacheKey, cancellationToken);
         if (cachedResponse != null)
         {
-            response = JsonConvert.DeserializeObject<TResponse>(Encoding.Default.GetString(cachedResponse));
+            response = JsonSerializer.Deserialize<TResponse>(Encoding.Default.GetString(cachedResponse))!;
             _logger.LogInformation($"Fetched from Cache -> {request.CacheKey}");
         }
         else
         {
-            response = await GetResponseAndAddToCache();
-            _logger.LogInformation($"Added to Cache -> {request.CacheKey}");
+            response = await getResponseAndAddToCache(request, next, cancellationToken);
         }
 
         return response;
     }
+
+    private async Task<TResponse> getResponseAndAddToCache(
+        TRequest request,
+        RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken
+    )
+    {
+        TResponse response = await next();
+
+        TimeSpan slidingExpiration = request.SlidingExpiration ?? TimeSpan.FromDays(_cacheSettings.SlidingExpiration);
+        DistributedCacheEntryOptions cacheOptions = new() { SlidingExpiration = slidingExpiration };
+
+        byte[] serializeData = Encoding.UTF8.GetBytes(JsonSerializer.Serialize(response));
+        await _cache.SetAsync(request.CacheKey, serializeData, cacheOptions, cancellationToken);
+        _logger.LogInformation($"Added to Cache -> {request.CacheKey}");
+
+        if (request.CacheGroupKey != null)
+            await addCacheKeyToGroup(request, slidingExpiration, cancellationToken);
+
+        return response;
+    }
+
+    private async Task addCacheKeyToGroup(TRequest request, TimeSpan slidingExpiration, CancellationToken cancellationToken)
+    {
+        byte[]? cacheGroupCache = await _cache.GetAsync(key: request.CacheGroupKey!, cancellationToken);
+        HashSet<string> cacheKeysInGroup;
+        if (cacheGroupCache != null)
+        {
+            cacheKeysInGroup = JsonSerializer.Deserialize<HashSet<string>>(Encoding.Default.GetString(cacheGroupCache))!;
+            if (!cacheKeysInGroup.Contains(request.CacheKey))
+                cacheKeysInGroup.Add(request.CacheKey);
+        }
+        else
+            cacheKeysInGroup = new HashSet<string>(new[] { request.CacheKey });
+        byte[] newCacheGroupCache = JsonSerializer.SerializeToUtf8Bytes(cacheKeysInGroup);
+
+        byte[]? cacheGroupCacheSlidingExpirationCache = await _cache.GetAsync(
+            key: $"{request.CacheGroupKey}SlidingExpiration",
+            cancellationToken
+        );
+        int? cacheGroupCacheSlidingExpirationValue = null;
+        if (cacheGroupCacheSlidingExpirationCache != null)
+            cacheGroupCacheSlidingExpirationValue = Convert.ToInt32(Encoding.Default.GetString(cacheGroupCacheSlidingExpirationCache));
+        if (cacheGroupCacheSlidingExpirationValue == null || slidingExpiration.TotalSeconds > cacheGroupCacheSlidingExpirationValue)
+            cacheGroupCacheSlidingExpirationValue = Convert.ToInt32(slidingExpiration.TotalSeconds);
+        byte[] serializeCachedGroupSlidingExpirationData = JsonSerializer.SerializeToUtf8Bytes(cacheGroupCacheSlidingExpirationValue);
+
+        DistributedCacheEntryOptions cacheOptions =
+            new() { SlidingExpiration = TimeSpan.FromSeconds(Convert.ToDouble(cacheGroupCacheSlidingExpirationValue)) };
+
+        await _cache.SetAsync(key: request.CacheGroupKey!, newCacheGroupCache, cacheOptions, cancellationToken);
+        _logger.LogInformation($"Added to Cache -> {request.CacheGroupKey}");
+
+        await _cache.SetAsync(
+            key: $"{request.CacheGroupKey}SlidingExpiration",
+            serializeCachedGroupSlidingExpirationData,
+            cacheOptions,
+            cancellationToken
+        );
+        _logger.LogInformation($"Added to Cache -> {request.CacheGroupKey}SlidingExpiration");
+    }
 }
+
+    
