@@ -2,45 +2,39 @@
 using Core.CrossCuttingConcerns.Logging.Serilog;
 using MediatR;
 using Microsoft.AspNetCore.Http;
-using Newtonsoft.Json;
+using System.Text.Json;
 
 namespace Core.Application.Pipelines.Logging;
 
 public class LoggingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
     where TRequest : IRequest<TResponse>, ILoggableRequest
 {
-    private readonly LoggerServiceBase _loggerServiceBase;
     private readonly IHttpContextAccessor _httpContextAccessor;
+    private readonly LoggerServiceBase _loggerServiceBase;
 
-
-    public LoggingBehavior(LoggerServiceBase loggerServiceBase, IHttpContextAccessor httpContextAccessor)
+    public LoggingBehavior(IHttpContextAccessor httpContextAccessor, LoggerServiceBase loggerServiceBase)
     {
-        _loggerServiceBase = loggerServiceBase;
         _httpContextAccessor = httpContextAccessor;
+        _loggerServiceBase = loggerServiceBase;
     }
 
-    public Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken,
-                                  RequestHandlerDelegate<TResponse> next)
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
     {
-        List<LogParameter> logParameters = new();
-        logParameters.Add(new LogParameter
-        {
-            Type = request.GetType().Name,
-            Value = request
-        });
+        List<LogParameter> logParameters =
+            new()
+            {
+                new LogParameter {Type=request.GetType().Name, Value=request}
+            };
 
-        LogDetail logDetail = new()
-        {
-            MethodName = next.Method.Name,
-            Parameters = logParameters,
-            User = _httpContextAccessor.HttpContext == null ||
-                   _httpContextAccessor.HttpContext.User.Identity.Name == null
-                       ? "?"
-                       : _httpContextAccessor.HttpContext.User.Identity.Name
-        };
+        LogDetail logDetail =
+            new()
+            {
+                MethodName=next.Method.Name,
+                Parameters=logParameters,
+                User=_httpContextAccessor.HttpContext.User.Identity?.Name ?? "?"
+            };
 
-        _loggerServiceBase.Info(JsonConvert.SerializeObject(logDetail));
-
-        return next();
+        _loggerServiceBase.Info(JsonSerializer.Serialize(logDetail));
+        return await next();
     }
 }
