@@ -3,5 +3,6 @@
 public interface ICacheRemoverRequest
 {
     bool BypassCache { get; }
-    string CacheKey { get; }
+    string? CacheKey { get; }
+    string? CacheGroupKey { get; }
 }
